+++ src/main/java/back/ahwhew/controller/UserController.java
@@ -13,6 +13,9 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
 
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 @RestController
 @Slf4j
 @RequestMapping("/auth")
@@ -38,26 +41,31 @@ public class UserController {
     @PostMapping("/signup")
     public ResponseEntity<?> registerUser(@RequestBody UserDTO dto) {
         try{
-            UserEntity user = UserEntity.builder()
-                    .userId(dto.getUserId())
-//                    .password(passwordEncoder.encode(dto.getPassword()))
-                    .password(dto.getPassword())
-                    .nickname(dto.getNickname())
-                    .age(dto.getAge())
-                    .gender(dto.getGender())
-                    .build();
+            log.info("Start signup");
 
-            UserEntity registeredUser = service.create(user);
+            if(SpecialCharacterCheck(dto.getPassword())) {
+                throw new RuntimeException("Password is invalid arguments");
+            }
 
-            UserDTO resDTO = UserDTO.builder()
-                    .userId(registeredUser.getUserId())
-                    .password(registeredUser.getPassword())
-                    .age(registeredUser.getAge())
-                    .nickname(registeredUser.getNickname())
-                    .gender(registeredUser.getGender())
-                    .build();
+                UserEntity user = UserEntity.builder()
+                        .userId(dto.getUserId())
+                        .password(passwordEncoder.encode(dto.getPassword()))
+                        .nickname(dto.getNickname())
+                        .age(dto.getAge())
+                        .gender(dto.getGender())
+                        .build();
 
-            return ResponseEntity.ok().body(resDTO);
+                UserEntity registeredUser = service.create(user);
+
+                UserDTO resDTO = UserDTO.builder()
+                        .userId(registeredUser.getUserId())
+                        .password(registeredUser.getPassword())
+                        .age(registeredUser.getAge())
+                        .nickname(registeredUser.getNickname())
+                        .gender(registeredUser.getGender())
+                        .build();
+
+                return ResponseEntity.ok().body(resDTO);
         } catch(Exception e) {
             ResponseDTO resDTO = ResponseDTO.builder().error(e.getMessage()).build();
             return ResponseEntity.badRequest().body(resDTO);
@@ -111,6 +119,16 @@ public class UserController {
             log.info("user vail checked");
             return true;
         }
+    }
 
+    // 특수 문자 확인
+    public Boolean SpecialCharacterCheck(String pw) {
+        // 확인해야 할 특수문자
+        String specialCharacterPattern = "[!@#$%^&*()\\\\?/.,]";
+
+        Pattern pattern = Pattern.compile(specialCharacterPattern);
+        Matcher matcher = pattern.matcher(pw);
+
+        return matcher.find();
     }
 }
